Create PeoplePresenceShutdown
This Module will shut a system down based on people presence. This is especially helpful if you have an always active source plugged into the codec. It will shut the system down if no one is in the room even if something was left sharing. The standby delay interval is set by the Standby Delay Configuration Setting.